Refactoring never stops
