+++ apis/betterself/v1/events/tests/views/test_views.py
@@ -6,81 +6,15 @@ from django.test import TestCase
 from django.urls import reverse
 from rest_framework.test import APIClient
 
-from apis.betterself.v1.signup.fixtures.builders import DemoHistoricalDataBuilder
 from apis.betterself.v1.tests.mixins.test_get_requests import GetRequestsTestsMixin
 from apis.betterself.v1.tests.mixins.test_post_requests import PostRequestsTestsMixin
-from apis.betterself.v1.tests.mixins.test_put_requests import PUTRequestsTestsMixin
 from apis.betterself.v1.tests.test_base import BaseAPIv1Tests
 from betterself.utils.date_utils import UTC_TZ, get_current_utc_time_and_tz
-from events.fixtures.factories import UserActivityFactory, UserActivityEventFactory
-from events.fixtures.mixins import UserActivityEventFixturesGenerator, UserSupplementStackFixturesGenerator
-from events.models import SupplementLog, UserActivity, UserActivityLog, SleepLog, \
-    SupplementReminder
-from supplements.models import Supplement, UserSupplementStack
+from events.fixtures.mixins import UserSupplementStackFixturesGenerator
+from events.models import SupplementLog, SleepLog
+from supplements.models import UserSupplementStack
 
 User = get_user_model()
-
-
-class TestUserActivityViews(BaseAPIv1Tests, GetRequestsTestsMixin, PostRequestsTestsMixin, PUTRequestsTestsMixin):
-    TEST_MODEL = UserActivity
-    PAGINATION = True
-    DEFAULT_POST_PARAMS = {
-        'name': 'Went For A Run',
-        'is_significant_activity': True
-    }
-    SECONDARY_ACTIVITY_NAME = 'Had Breakfast'
-    SECONDARY_POST_PARAMS = {
-        'name': SECONDARY_ACTIVITY_NAME,
-        'is_significant_activity': False
-    }
-
-    @classmethod
-    def setUpTestData(cls):
-        super().setUpTestData()
-        UserActivityEventFixturesGenerator.create_fixtures(cls.user_1)
-        UserActivityFactory(user=cls.user_1, **cls.SECONDARY_POST_PARAMS)
-
-    def test_valid_get_request_for_key_in_response(self):
-        key = 'name'
-        super().test_valid_get_request_for_key_in_response(key)
-
-    def test_valid_get_request_with_params_filters_correctly(self):
-        request_parameters = {'name': self.SECONDARY_ACTIVITY_NAME}
-        super().test_valid_get_request_with_params_filters_correctly(request_parameters)
-
-
-class TestUserActivityEventViews(BaseAPIv1Tests, GetRequestsTestsMixin, PostRequestsTestsMixin, PUTRequestsTestsMixin):
-    TEST_MODEL = UserActivityLog
-    PAGINATION = True
-
-    def setUp(self):
-        current_time = datetime.datetime.now()
-        previous_time = current_time - relativedelta(days=5)
-        self.DEFAULT_POST_PARAMS = {
-            'time': previous_time.isoformat(),
-            'source': 'user_excel',
-            'duration_minutes': 30
-        }
-
-        # pass a parameter just to make sure the default parameter is valid
-        self.valid_user_activity = UserActivity.objects.filter(user=self.user_1).first()
-        self.DEFAULT_POST_PARAMS['user_activity_uuid'] = str(self.valid_user_activity.uuid)
-
-        super().setUp()
-
-    @classmethod
-    def setUpTestData(cls):
-        super().setUpTestData()
-        UserActivityEventFixturesGenerator.create_fixtures(cls.user_1)
-
-    def test_valid_get_request_for_key_in_response(self):
-        key = 'duration_minutes'
-        super().test_valid_get_request_for_key_in_response(key)
-
-    def test_valid_get_request_with_params_filters_correctly(self):
-        UserActivityEventFactory(user=self.user_1, duration_minutes=30)
-        request_parameters = {'duration_minutes': 30}
-        super().test_valid_get_request_with_params_filters_correctly(request_parameters)
 
 
 class TestSleepActivityViews(BaseAPIv1Tests, GetRequestsTestsMixin, PostRequestsTestsMixin):
@@ -197,72 +131,6 @@ class TestSleepActivityViews(BaseAPIv1Tests, GetRequestsTestsMixin, PostRequests
 
         response = self._make_post_request(self.client_1, request_parameters)
         self.assertEqual(response.status_code, 400, response.data)
-
-
-class TestSupplementReminderViews(BaseAPIv1Tests, PostRequestsTestsMixin):
-    TEST_MODEL = SupplementReminder
-    PAGINATION = False
-
-    @classmethod
-    def setUpTestData(cls):
-        cls.user_1, _ = User.objects.get_or_create(username='default')
-
-        builder = DemoHistoricalDataBuilder(cls.user_1)
-        builder.create_historical_fixtures()
-        builder.create_supplement_reminders(limit=4)
-
-        cls.url = reverse(SupplementReminder.RESOURCE_NAME)
-        super().setUpTestData()
-
-    def setUp(self):
-        supplement = Supplement.objects.filter(user=self.user_1).first()
-        supplement_uuid = str(supplement.uuid)
-
-        self.DEFAULT_POST_PARAMS = {
-            'reminder_time': '15:20',
-            'quantity': 5,
-            'supplement_uuid': supplement_uuid
-        }
-
-        self.client_1 = self.create_authenticated_user_on_client(APIClient(), self.user_1)
-        self.client_2 = self.create_authenticated_user_on_client(APIClient(), self.user_2)
-
-    def test_post_when_over_limit(self):
-        # hardcoded value of 5 to prevent spam
-        supplements = Supplement.objects.filter(user=self.user_1)
-        for supplement in supplements:
-            params = {
-                'reminder_time': '10:20',
-                'quantity': 5,
-                'supplement_uuid': str(supplement.uuid)
-            }
-            self.client_1.post(self.url, data=params)
-
-        cutoff_limit = 5
-        user_supplement_reminders = SupplementReminder.objects.filter(user=self.user_1).count()
-        self.assertEqual(cutoff_limit, user_supplement_reminders)
-
-    def test_view_no_auth(self):
-        client = APIClient()
-        response = client.get(self.url)
-        self.assertEqual(response.status_code, 401)
-
-    def test_view_no_data(self):
-        new_user, _ = User.objects.get_or_create(username='no-data')
-
-        client = APIClient()
-        client.force_login(new_user)
-
-        response = client.get(self.url)
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(len(response.data), 0)
-
-    def test_view(self):
-        response = self.client_1.get(self.url)
-        self.assertEqual(response.status_code, 200)
-
-        supplement_reminder_count = SupplementReminder.objects.filter(user=self.user_1).count()
-        self.assertEqual(supplement_reminder_count, len(response.data))
 
 
 class TestUserStackRecordView(TestCase):
